refactor: remove 'project.project' dependencies
Remove name redundancies when declaring imports from project folder.
