+++ src/encord_active/lib/db/predictions.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel, Field, validator
 from encord_active.lib.metrics.execute import run_all_prediction_metrics
 from encord_active.lib.model_predictions.iterator import PredictionIterator
 from encord_active.lib.model_predictions.writer import PredictionWriter
-from encord_active.lib.project.project import Project
+from encord_active.lib.project import Project
 
 RelativeFloat = Annotated[float, Field(ge=0, le=1)]
