+++ src/encord_active/lib/model_predictions/writer.py
@@ -14,7 +14,7 @@ from torchvision.ops import box_iou
 from tqdm import tqdm
 
 from encord_active.lib.common.utils import binary_mask_to_rle, rle_iou
-from encord_active.lib.project.project import Project
+from encord_active.lib.project import Project
 
 logger = logging.getLogger(__name__)
 BBOX_KEYS = {"x", "y", "w", "h"}
