+++ src/encord_active/lib/project/__init__.py
@@ -1,0 +1,2 @@
+from .project import Project
+from .project_file_structure import LabelRowStructure, ProjectFileStructure
